# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Solar API",
    "description": "API interne pour la plateforme Solar (Association Le Trèfle 2.0). Supporte l'authentification par Bearer JWT ou API Key.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.beta.letrefle.org/",
      "description": "Serveur de production"
    },
    {
      "url": "http://localhost:3001",
      "description": "Serveur de développement local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "name": "token",
        "in": "header"
      },
      "apiKeyQuery": {
        "type": "apiKey",
        "name": "token",
        "in": "query"
      }
    },
    "schemas": {
...
      "Newsletter": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "scheduled",
              "sent"
            ]
          },
          "authorId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "scheduledAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyAuth": []
    },
    {
      "apiKeyQuery": []
    }
  ],
  "paths": {
...
    "/v1/newsletters": {
      "get": {
        "summary": "Lister les newsletters",
        "tags": [
          "Newsletters"
        ],
        "responses": {
          "200": {
            "description": "Liste des newsletters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Newsletter"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Créer une newsletter",
        "tags": [
          "Newsletters"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "htmlContent": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Newsletter créée"
          }
        }
      }
    },
...
    "/v1/newsletters/{id}": {
      "get": {
        "summary": "Récupérer une newsletter",
        "tags": [
          "Newsletters"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Détails de la newsletter"
          }
        }
      },
      "put": {
        "summary": "Mettre à jour une newsletter",
        "tags": [
          "Newsletters"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "htmlContent": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "draft",
                      "scheduled"
                    ]
                  },
                  "scheduledAt": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mise à jour réussie"
          }
        }
      },
      "delete": {
        "summary": "Supprimer une newsletter",
        "tags": [
          "Newsletters"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Newsletter supprimée"
          }
        }
      }
    },
    "/v1/newsletters/{id}/unschedule": {
      "post": {
        "summary": "Annuler la planification d'une newsletter",
        "tags": [
          "Newsletters"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Planification annulée"
          }
        }
      }
    },
    "/v1/newsletters/{id}/send": {
      "post": {
        "summary": "Envoyer une newsletter immédiatement",
        "tags": [
          "Newsletters"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Newsletter envoyée"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "https://api.beta.letrefle.org/"
    }
  ]
}